+++ apps/common/views.py
@@ -1,0 +1,5 @@
+from django.http import HttpResponse
+from django.shortcuts import render
+from datetime import datetime
+
+# Create your views here.