+++ apps/pages/models.py
@@ -1,0 +1,4 @@
+from django.db import models
+from autoslug import AutoSlugField
+
+# Create your models here.