+++ core/urls.py
@@ -20,7 +20,7 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("", include('home.urls')),
+    path("", include('apps.pages.urls')),
     path("users/", include("apps.users.urls")),
     path('', include('apps.blog.urls')),
     path('accounts/', include('allauth.urls')),